+++ lib/particlefx2d.rb
@@ -2,6 +2,7 @@
 
 require_relative 'particlefx2d/version'
 require_relative 'particlefx2d/emitter'
+require_relative 'particlefx2d/renderer_factory'
 
 # 2D particle effects API for use with any Ruby graphics API.
 module ParticleFX2D