+++ lib/particlefx_ruby2d.rb
@@ -3,6 +3,8 @@
 require_relative 'particlefx2d'
 require_relative 'particlefx2d/ruby2d/particle_circle'
 require_relative 'particlefx2d/ruby2d/particle_image'
+require_relative 'particlefx2d/ruby2d/shape_renderer_factory'
+require_relative 'particlefx2d/ruby2d/canvas_renderer_factory'
 
 module ParticleFX2D
   # This module defines the Ruby2D-specific particle shapes